Commit - move file from dev api to package data.api, and change class api config to object
