--- app/src/main/java/com/dicoding/smartbin/ApiConfig.kt
@@ -1,25 +1,0 @@
-package com.dicoding.smartbin
-
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class ApiConfig {
-    companion object{
-        fun getApiService(): ApiService{
-            val loggingInterceptor =
-                HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY)
-            val client = OkHttpClient.Builder()
-                .addInterceptor(loggingInterceptor)
-                .build()
-
-            val retrofit = Retrofit.Builder()
-                .baseUrl("")//di isi sama IP internet recommend dari hotspot
-                .addConverterFactory(GsonConverterFactory.create())
-                .client(client)
-                .build()
-            return retrofit.create(ApiService::class.java)
-        }
-    }
-}
